+++ plugin_ide.ui/src-dltk-branch/_org/eclipse/dltk/internal/ui/text/ScriptReconciler.java
@@ -24,9 +24,11 @@ import org.eclipse.dltk.core.ElementChangedEvent;
 import org.eclipse.dltk.core.IElementChangedListener;
 import org.eclipse.dltk.core.IModelElementDelta;
 import org.eclipse.dltk.core.ISourceModule;
-import org.eclipse.dltk.internal.ui.editor.EditorUtility;
 import org.eclipse.dltk.internal.ui.editor.ScriptEditor;
+
+import _org.eclipse.dltk.internal.ui.editor.EditorUtility;
 import _org.eclipse.dltk.internal.ui.text.ScriptCompositeReconcilingStrategy;
+
 import org.eclipse.jface.preference.IPreferenceStore;
 import org.eclipse.jface.text.IDocument;
 import org.eclipse.jface.text.ITextViewer;