+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/text/completion/LangContentAssistProcessor.java
@@ -98,6 +98,7 @@ public class LangContentAssistProcessor extends ContenAssistProcessorExt {
 		protected ILangCompletionProposalComputer createSnippetsProposalComputer() {
 			return new LangTemplateCompletionProposalComputer();
 		}
+		
 	}
 	
 	/* -----------------  ----------------- */